refactor: 패키지명을 domain과 view로 변경
domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현
