+++ src/main/kotlin/study/step3/domain/Mileage.kt
@@ -1,4 +1,4 @@
-package study.step3.entity
+package study.step3.domain
 
 @JvmInline
 value class Mileage(val value: Int) {
